requirements.txt
mysqlclient ~= 2.1.0
SQLAlchemy ~= 1.4.31
apache-airflow-providers-docker
pandas
numpy